--- packages/components/form/src/form.type.ts
@@ -1,9 +1,0 @@
-import type { RuleItem } from 'async-validator'
-
-export interface FormItemRule extends RuleItem {
-  trigger?: string
-}
-
-export type FormRulesMap<T extends string = string> = Partial<
-  Record<T, FormItemRule | FormItemRule[]>
->
